[BOOK-491] feat: 갤러리 이미지 연동 및 OCR 화면 구조 리팩토링
- OCR 화면의 UI 상태에 따라 `OcrCameraView`, `OcrImageView`, `OcrResultView`로 뷰를 분리하여 구조를 개선
