+++ feature/record/src/main/kotlin/com/ninecraft/booket/feature/record/ocr/OcrUi.kt
@@ -7,6 +7,7 @@ import android.provider.Settings
 import android.view.ViewGroup
 import android.widget.LinearLayout
 import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.PickVisualMediaRequest
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.camera.core.ImageCapture
 import androidx.camera.core.ImageCaptureException
@@ -48,6 +49,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import com.ninecraft.booket.core.designsystem.ComponentPreview
+import com.ninecraft.booket.core.designsystem.component.NetworkImage
 import com.ninecraft.booket.core.designsystem.component.button.ReedButton
 import com.ninecraft.booket.core.designsystem.component.button.ReedButtonColorStyle
 import com.ninecraft.booket.core.designsystem.component.button.largeButtonStyle
@@ -61,6 +63,9 @@ import com.ninecraft.booket.feature.record.R
 import com.ninecraft.booket.feature.record.ocr.component.CameraBottomBar
 import com.ninecraft.booket.feature.record.ocr.component.CameraFrame
 import com.ninecraft.booket.feature.record.ocr.component.SentenceBox
+import com.ninecraft.booket.feature.record.ocr.view.OcrCameraView
+import com.ninecraft.booket.feature.record.ocr.view.OcrImageView
+import com.ninecraft.booket.feature.record.ocr.view.OcrResultView
 import com.ninecraft.booket.feature.screens.OcrScreen
 import com.skydoves.compose.stability.runtime.TraceRecomposition
 import com.slack.circuit.codegen.annotations.CircuitInject
@@ -78,339 +83,8 @@ internal fun OcrUi(
     HandleOcrSideEffects(state = state)
 
     when (state.currentUi) {
-        OcrUi.CAMERA -> CameraPreview(state = state, modifier = modifier)
-        OcrUi.RESULT -> TextScanResult(state = state, modifier = modifier)
-    }
-}
-
-@TraceRecomposition
-@Composable
-private fun CameraPreview(
-    state: OcrUiState,
-    modifier: Modifier = Modifier,
-) {
-    val context = LocalContext.current
-    val lifecycleOwner = LocalLifecycleOwner.current
-    val permission = android.Manifest.permission.CAMERA
-
-    /**
-     * Camera Permission Request
-     */
-    val isGranted by produceState(
-        initialValue = ContextCompat.checkSelfPermission(context, permission) == PackageManager.PERMISSION_GRANTED,
-        key1 = lifecycleOwner, // lifecycle 변경 시 재설정
-    ) {
-        // 최초 동기화
-        value = ContextCompat.checkSelfPermission(context, permission) == PackageManager.PERMISSION_GRANTED
-
-        // 포그라운드 복귀 시 OS 권한 동기화
-        val observer = LifecycleEventObserver { _, event ->
-            if (event == Lifecycle.Event.ON_RESUME) {
-                value = ContextCompat.checkSelfPermission(context, permission) == PackageManager.PERMISSION_GRANTED
-                if (value) {
-                    state.eventSink(OcrUiEvent.OnHidePermissionDialog)
-                } else {
-                    state.eventSink(OcrUiEvent.OnShowPermissionDialog)
-                }
-            }
-        }
-        lifecycleOwner.lifecycle.addObserver(observer)
-        awaitDispose { lifecycleOwner.lifecycle.removeObserver(observer) }
-    }
-
-    val permissionLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.RequestPermission(),
-    ) { granted ->
-        if (!granted) {
-            state.eventSink(OcrUiEvent.OnShowPermissionDialog)
-        }
-    }
-    val settingsLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.StartActivityForResult(),
-    ) { _ -> }
-
-    // 최초 진입 시 권한 요청
-    LaunchedEffect(Unit) {
-        if (!isGranted) {
-            state.eventSink(OcrUiEvent.OnHidePermissionDialog)
-            permissionLauncher.launch(permission)
-        }
-    }
-
-    /**
-     * Camera Controller
-     */
-    val cameraController = remember { LifecycleCameraController(context) }
-
-    DisposableEffect(isGranted, lifecycleOwner, cameraController) {
-        if (isGranted) {
-            cameraController.bindToLifecycle(lifecycleOwner)
-        }
-
-        onDispose {
-            cameraController.unbind()
-        }
-    }
-
-    /**
-     * SystemStatusBar Color
-     */
-    val systemUiController = rememberSystemUiController()
-
-    DisposableEffect(systemUiController) {
-        systemUiController.setSystemBarsColor(
-            color = Color.Transparent,
-            darkIcons = false,
-            isNavigationBarContrastEnforced = false,
-        )
-
-        onDispose {
-            systemUiController.setSystemBarsColor(
-                color = Color.Transparent,
-                darkIcons = true,
-                isNavigationBarContrastEnforced = false,
-            )
-        }
-    }
-
-    ReedScaffold(
-        modifier = modifier.fillMaxSize(),
-        containerColor = Neutral950,
-    ) { innerPadding ->
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(innerPadding),
-        ) {
-            ReedCloseTopAppBar(
-                modifier = Modifier
-                    .background(color = Color.Black)
-                    .align(Alignment.TopCenter),
-                isDark = true,
-                onClose = {
-                    state.eventSink(OcrUiEvent.OnCloseClick)
-                },
-            )
-            Text(
-                text = stringResource(R.string.ocr_guide),
-                modifier = Modifier
-                    .align(Alignment.Center)
-                    .offset {
-                        IntOffset(
-                            x = 0,
-                            y = (-164).dp.roundToPx(),
-                        )
-                    },
-                color = ReedTheme.colors.contentInverse,
-                textAlign = TextAlign.Center,
-                style = ReedTheme.typography.headline2Medium,
-            )
-
-            if (isGranted) {
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .background(color = White)
-                        .height(200.dp)
-                        .align(Alignment.Center),
-                ) {
-                    AndroidView(
-                        modifier = Modifier.fillMaxSize(),
-                        factory = { context ->
-                            PreviewView(context).apply {
-                                layoutParams = LinearLayout.LayoutParams(
-                                    ViewGroup.LayoutParams.MATCH_PARENT,
-                                    ViewGroup.LayoutParams.MATCH_PARENT,
-                                )
-                                clipToOutline = true
-                                implementationMode = PreviewView.ImplementationMode.COMPATIBLE
-                                scaleType = PreviewView.ScaleType.FILL_CENTER
-                                controller = cameraController
-                            }
-                        },
-                    )
-                }
-                CameraFrame(modifier = Modifier.align(Alignment.Center))
-            }
-
-            Column(
-                modifier = Modifier.align(Alignment.BottomCenter),
-                horizontalAlignment = Alignment.CenterHorizontally,
-            ) {
-                if (state.isTextDetectionFailed) {
-                    Text(
-                        text = stringResource(R.string.ocr_error_text_detection_failed),
-                        color = ReedTheme.colors.contentError,
-                        textAlign = TextAlign.Center,
-                        style = ReedTheme.typography.label2Regular,
-                    )
-                    Spacer(modifier = Modifier.height(22.dp))
-                }
-                CameraBottomBar(
-                    onGalleryClick = {
-                        // TODO: 갤러리
-                    },
-                    onCaptureClick = {
-                        state.eventSink(OcrUiEvent.OnCaptureStart)
-
-                        val executor = ContextCompat.getMainExecutor(context)
-                        val photoFile = File.createTempFile("ocr_", ".jpg", context.cacheDir)
-                        val output = ImageCapture.OutputFileOptions.Builder(photoFile).build()
-
-                        cameraController.takePicture(
-                            output, executor,
-                            object : ImageCapture.OnImageSavedCallback {
-                                override fun onImageSaved(outputFileResults: ImageCapture.OutputFileResults) {
-                                    state.eventSink(OcrUiEvent.OnImageCaptured(photoFile.toUri()))
-                                }
-
-                                override fun onError(exception: ImageCaptureException) {
-                                    state.eventSink(OcrUiEvent.OnCaptureFailed(exception))
-                                }
-                            },
-                        )
-                    },
-                    buttonEnabled = !state.isLoading,
-                )
-
-                Spacer(modifier = Modifier.height(ReedTheme.spacing.spacing4))
-            }
-
-            if (state.isLoading) {
-                Box(
-                    modifier = Modifier.fillMaxSize(),
-                    contentAlignment = Alignment.Center,
-                ) {
-                    CircularProgressIndicator(color = ReedTheme.colors.contentBrand)
-                }
-            }
-        }
-    }
-
-    if (state.isPermissionDialogVisible) {
-        ReedDialog(
-            title = stringResource(R.string.permission_dialog_title),
-            description = stringResource(R.string.permission_dialog_description),
-            confirmButtonText = stringResource(R.string.permission_dialog_move_to_settings),
-            onConfirmRequest = {
-                val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS).apply {
-                    data = Uri.fromParts("package", context.packageName, null)
-                }
-                settingsLauncher.launch(intent)
-            },
-        )
-    }
-}
-
-@TraceRecomposition
-@Composable
-private fun TextScanResult(
-    state: OcrUiState,
-    modifier: Modifier = Modifier,
-) {
-    ReedScaffold(
-        modifier = modifier.fillMaxSize(),
-        containerColor = White,
-    ) { innerPadding ->
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(innerPadding),
-        ) {
-            ReedCloseTopAppBar(
-                title = stringResource(R.string.ocr_sentence_selection),
-                onClose = {
-                    state.eventSink(OcrUiEvent.OnCloseClick)
-                },
-            )
-            LazyColumn(
-                modifier = Modifier
-                    .weight(1f)
-                    .padding(horizontal = ReedTheme.spacing.spacing5),
-                verticalArrangement = Arrangement.spacedBy(ReedTheme.spacing.spacing2),
-            ) {
-                item {
-                    Spacer(modifier = Modifier.height(ReedTheme.spacing.spacing1))
-                }
-
-                items(state.sentenceList.size) { index ->
-                    SentenceBox(
-                        onClick = {
-                            state.eventSink(OcrUiEvent.OnSentenceSelected(index))
-                        },
-                        sentence = state.sentenceList[index],
-                        isSelected = state.selectedIndices.contains(index),
-                    )
-                }
-            }
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(
-                        horizontal = ReedTheme.spacing.spacing5,
-                        vertical = ReedTheme.spacing.spacing4,
-                    ),
-            ) {
-                ReedButton(
-                    onClick = {
-                        state.eventSink(OcrUiEvent.OnReCaptureButtonClick)
-                    },
-                    sizeStyle = largeButtonStyle,
-                    colorStyle = ReedButtonColorStyle.SECONDARY,
-                    modifier = Modifier.weight(1f),
-                    text = stringResource(R.string.ocr_recapture),
-                )
-                Spacer(modifier = Modifier.width(ReedTheme.spacing.spacing2))
-                ReedButton(
-                    onClick = {
-                        state.eventSink(OcrUiEvent.OnSelectionConfirmed)
-                    },
-                    sizeStyle = largeButtonStyle,
-                    colorStyle = ReedButtonColorStyle.PRIMARY,
-                    enabled = state.selectedIndices.isNotEmpty(),
-                    modifier = Modifier.weight(1f),
-                    text = stringResource(R.string.ocr_selection_confirm),
-                )
-            }
-        }
-    }
-
-    if (state.isRecaptureDialogVisible) {
-        ReedDialog(
-            title = stringResource(R.string.recapture_dialog_title),
-            description = stringResource(R.string.recapture_dialog_description),
-            confirmButtonText = stringResource(R.string.recapture_dialog_confirm),
-            onConfirmRequest = {
-                state.eventSink(OcrUiEvent.OnRecaptureDialogConfirmed)
-            },
-            dismissButtonText = stringResource(R.string.recapture_dialog_cancel),
-            onDismissRequest = {
-                state.eventSink(OcrUiEvent.OnRecaptureDialogDismissed)
-            },
-        )
-    }
-}
-
-@ComponentPreview
-@Composable
-private fun CameraPreviewPreview() {
-    ReedTheme {
-        CameraPreview(
-            state = OcrUiState(
-                eventSink = {},
-            ),
-        )
-    }
-}
-
-@ComponentPreview
-@Composable
-private fun TextRecognitionResultPreview() {
-    ReedTheme {
-        TextScanResult(
-            state = OcrUiState(
-                eventSink = {},
-            ),
-        )
+        OcrUi.CAMERA -> OcrCameraView(state = state, modifier = modifier)
+        OcrUi.IMAGE -> OcrImageView(state = state, modifier = modifier)
+        OcrUi.RESULT -> OcrResultView(state = state, modifier = modifier)
     }
 }
